#554: move motive-coverage check into PkMonoTest
jtcop expects each test class to have a same-named production
class (RuleEveryTestHasProductionClass) and present-tense test
names (RulePresentTense). Fold the new motive-coverage assertion
into PkMonoTest as linksEveryMotiveFileToALint() and drop the
standalone MotivesAreUsedTest.
